lambda_package_processer/requirements.txt
pydantic==2.11.7
loguru==0.7.3
numpy==1.26.4
pandas==2.2.2
pydantic-settings
loguru
motor
odmantic
pymongo
google-generativeai==0.8.5
google-cloud-aiplatform==1.95.1
pyyaml
boto3